RUM-2925 Move `DDCrashReport` APIs to `DatadogInternal`
