+++ DatadogCrashReporting/Sources/ThirdPartyCrashReporter.swift
@@ -4,7 +4,7 @@
  * Copyright 2019-Present Datadog, Inc.
  */
 
-import Foundation
+import DatadogInternal
 
 /// An interface of 3rd party crash reporter used by the DatadogCrashReporting.
 internal protocol ThirdPartyCrashReporter {
