+++ DatadogCrashReporting/Tests/Mocks.swift
@@ -4,6 +4,7 @@
  * Copyright 2019-Present Datadog, Inc.
  */
 
+import DatadogInternal
 import CrashReporter
 
 @testable import DatadogCrashReporting
